[M] Move FormatUtils to Table, implement page presenter
